+++ src/clientintrospect.cpp
@@ -83,7 +83,7 @@ struct InfoOp : public OperationBase
         {
             (void)evbuffer_drain(conn->txBody.get(), evbuffer_get_length(conn->txBody.get()));
 
-            EvOutBuf R(hostBE, conn->txBody.get());
+            EvOutBuf R(conn->sendBE, conn->txBody.get());
 
             to_wire(R, chan->sid);
             to_wire(R, ioid);