+++ src/serverintrospect.cpp
@@ -35,7 +35,7 @@ struct ServerIntrospect : public ServerOp
         {
             (void)evbuffer_drain(conn->txBody.get(), evbuffer_get_length(conn->txBody.get()));
 
-            EvOutBuf R(hostBE, conn->txBody.get());
+            EvOutBuf R(conn->sendBE, conn->txBody.get());
             to_wire(R, uint32_t(ioid));
             to_wire(R, sts);
             if(type)